// src/app/models/schedules/schedule.ts
export class Schedule {
    schedule_id: string;
    from_time: any;
    to_time: any;
    doctor_id: string;
    doctor_name: string;
    doctor_type: string;
    quota: number;
    reservation: number;
    walkin: number;
    hospital_id: string;
    hospital_name: string;
    floor_name: string;
    wing_name: string;
    room_name: string;
  }