// src/app/models/doctors/doctor-note.ts
export class DoctorNote {
    doctor_note_id: string;
    doctor_id: string;
    hospital_id: string;
    from_date: string;
    to_date: string;
    note: string;
    is_active: string;
    created_by: string;
    created_date: string;
    created_from: string;
    modified_by: string;
    modified_date: string;
    modified_from: string;
}