// src/app/models/doctors/doctor-profile.ts
export class DoctorProfile {
    doctor_hope_id: number;
    doctor_id: string;
    doctor_type_id: string;
    doctor_type_name: string;
    effective_date: string;
    hospital_id: string;
    hospital_name: string;
    name: string;
    quota: number;
    reservation: number;
    walkin: number;
    speciality_id: string;
    specialization_id: string;
    specialization_name: string;
    specialization_name_en: string;
}